make ramp corr a processing step, use newer framework
